+++ CHANGELOG.rst
@@ -57,6 +57,7 @@ Changed
     - Refactor sheet config helpers into ``SheetConfigAPI`` (#905)
     - Include top header in column width estimation for rendering (#649)
     - Use node header for recognizing unit enabled columns without data (#914)
+    - Prevent simultaneous editing of cells in multiple tables (#765)
 
 Fixed
 -----